refactor: WebUIChessApplication 기능 분리 및 리팩토링
